Fixed some portability issues

+++ code/chapter6/mtime.h
@@ -1,3 +1,10 @@
+#include <stdint.h>
+
+// Initialize the RISC-V machine timer
 void mtime_init();
-long long mtime_get();
-void mtime_reset(long long quantum);
+
+// Return the value of the current MTIME register, in microseconds
+uint64_t mtime_get();
+
+// Request a timer interrupt after the given number of microseconds
+void mtime_reset(uint64_t quantum);
